backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
google-genai
python-dotenv==1.0.1
pydantic==2.7.1
beautifulsoup4==4.12.3
requests==2.31.0
PyPDF2==3.0.1
python-multipart==0.0.9